+++ skrub/__init__.py
@@ -8,7 +8,7 @@ from ._datetime_encoder import DatetimeEncoder
 from ._deduplicate import compute_ngram_distance, deduplicate
 from ._fuzzy_join import fuzzy_join
 from ._gap_encoder import GapEncoder
-from ._interpolation_join import InterpolationJoin
+from ._interpolation_join import InterpolationJoiner
 from ._joiner import Joiner
 from ._minhash_encoder import MinHashEncoder
 from ._similarity_encoder import SimilarityEncoder
@@ -26,7 +26,7 @@ __all__ = [
     "Joiner",
     "fuzzy_join",
     "GapEncoder",
-    "InterpolationJoin",
+    "InterpolationJoiner",
     "MinHashEncoder",
     "SimilarityEncoder",
     "SuperVectorizer",